Moved colors into constants folder
