+++ src/components/header.tsx
@@ -2,7 +2,7 @@ import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import Link from "next/link";
 import styled from "styled-components";
 import { path } from "../utils/url/drello-web";
-import { drelloColors } from "./shared-styles";
+import { drelloColors } from "../constants/colors";
 
 const HeaderContainer = styled.header`
   display: grid;
