+++ src/components/shared-styles.ts
@@ -1,11 +1,5 @@
 import styled from "styled-components";
 
-export const drelloColors = {
-  black: (opacity: number = 1) => `rgba(0,0,0,${opacity})`,
-  white: (opacity: number = 1) => `rgba(255,255,255,${opacity})`,
-  greyish: (opacity: number = 1) => `rgba(141, 153, 174,${opacity})`,
-};
-
 // Main body
 export const Main = styled.main`
   text-align: center;
