+++ src/constants/colors.ts
@@ -1,0 +1,5 @@
+export const drelloColors = {
+  black: (opacity: number = 1) => `rgba(0,0,0,${opacity})`,
+  white: (opacity: number = 1) => `rgba(255,255,255,${opacity})`,
+  greyish: (opacity: number = 1) => `rgba(141, 153, 174,${opacity})`,
+};
